refactor for external gasless package
