+++ internal/core/basetx/transactors.go
@@ -1,0 +1,18 @@
+package basetx
+
+import (
+	"context"
+
+	"github.com/ethereum/go-ethereum/accounts/abi/bind"
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/core/types"
+	"github.com/ipfs/go-cid"
+)
+
+func (t *Transactor) CreateOrganizationTx(ctx context.Context, txopts *bind.TransactOpts, metaCID cid.Cid) (*types.Transaction, error) {
+	return t.valist.CreateOrganization(txopts, metaCID.String())
+}
+
+func (t *Transactor) LinkOrganizationNameTx(ctx context.Context, txopts *bind.TransactOpts, orgID common.Hash, name string) (*types.Transaction, error) {
+	return t.registry.LinkNameToID(txopts, orgID, name)
+}
